requirements.txt
cmake==3.17.2
dlib==19.18.0
Pillow==9.2.0
numpy==1.14.5
pandas==1.5.2
gdown==3.10.1
tqdm==4.30.0
opencv-python==4.6.0.66
tensorflow==2.11.0
keras==2.2.0
Flask==1.1.2
matplotlib==3.2.2
deepface==0.0.75
mtcnn==0.1.1


